Rework to the 'pen name' topic.
This was a suggestion on the forums about a month ago. Um...better late than never, right? 

Let it be known that I wasn't entirely sure what I was doing/going off the normal name change dialogue to make sure I was doing it right. I'm also still spotty on how to use the new expressions. Those are my biggest concerns with this submission right now, but I gave it my all.